General refactor...   Express lives in the API folder
