+++ test/events.test.js
@@ -4,7 +4,7 @@ const { expect } = require('chai')
 const chaiHttp = require('chai-http')
 const faker = require('faker')
 chai.use(chaiHttp)
-const application = require('../app')
+const application = require('../api/app')
 const app = chai.request.agent(application).keepOpen()
 
 describe('#Events test', () => {
